requirements.txt
dj-database-url==0.4.2
Django==2.0.1
gunicorn==19.7.1
psycopg2==2.7.4
python-decouple==3.1
pytz==2017.3
yapf==0.20.2
